Move doc_cfg-specific code into `cfg.rs`
